Adding two more story boards
Added a page for which body part the user selects and one to enter their height and weight. I deleted a navigation controller I added in a previous commit because it made it to where there was no "back" button on one of the pages. For both of these pages, I created view controllers too.